(*) Optimizations

+++ src/driver.h
@@ -124,7 +124,7 @@ FCEUGI *FCEUI_CopyFamiStart();
 int FCEUI_Initialize(void);
 
 /* Emulates a frame. */
-void FCEUI_Emulate(uint8 **, int32 **, int32 *, int);
+void FCEUI_Emulate(uint8 **, int16 **, int32 *, int);
 
 /* Closes currently loaded game */
 void FCEUI_CloseGame(void);
